Improve the PollingMasterInquireClient logic
### What changes are proposed in this pull request?

Previously the service type in polling master inquire client was deduced
by looking up the master ip address. This is an unnecessary step because
when the client is initialized, the caller knows what exact type of
master it talks too. So we add this as a constructor param and let every
caller to pass it in, to improve the code readability.

### Why are the changes needed?

Please clarify why the changes are needed. For instance,
To improve the code readability.

### Does this PR introduce any user facing changes?

Please list the user-facing changes introduced by your change, including
N/A

pr-link: Alluxio/alluxio#16588
change-id: cid-ff9451fb78113580c8c871937d94649643d17b18