Add more observability on inode tree corruption
### What changes are proposed in this pull request?

Add more logs to detect inode tree corruption

### Why are the changes needed?

Some users reported a issue that expecting to lock inode A but inode B
is locked. After the investigation, we figured that this might be due to
inode tree corruption caused by journal loss. So we add more logs to
help users debug the issue better.

### Does this PR introduce any user facing changes?

N/A

pr-link: Alluxio/alluxio#16710
change-id: cid-b90490c7122465ccaaf83c660e8e5dfcfc0f692a