Fix bug for ufs journal dumper when read regular checkpoint
### What changes are proposed in this pull request?

Fix `UfsJournalDumper.dumpJournal()`, which shouldn't call
`Files.createDirectories(dir)`.

### Why are the changes needed?

bug: when using UfsJournal, if you use JournalTool to dump the ufs
journal of a regular checkpoint(like BlockMaster), it will throws
FileNotFoundException: /path-to-output/checkpoints-106 (Is a directory).
like this:
<img width="944" alt="截屏2022-11-15 11 15 58"
src="https://user-images.githubusercontent.com/5937168/201818808-9264ffb6-f5ff-491c-9a52-cbdada78bb8c.png">

Because in `AbstractJournalDumper.readCheckpoint(checkpoint, path)`,
path is a file or dir. For regular checkpoint, it is a file; and for
compound checkpoint, it is a dir. But `UfsJournalDumper` calls
`Files.createDirectories(dir)`, which should be called in
`AbstractJournalDumper.readCompoundCheckpoint`.

### Does this PR introduce any user facing changes?

None

pr-link: Alluxio/alluxio#16550
change-id: cid-19f14a9065879b141a5e0301592e1d5827d96a3d