Use path's hash as PageId of PagingCache on worker and client
PageId of PagingCache is used as directory name on local file system.
This is expected to be a valid file name string.
Apparently a Alluxio or UFS path may contain some special charactors.
Also it may exceed file name length limit.
So, we introduce a new hash function to generate such string as dir
name. Currently SHA256 is preferred as its hash function.

Signed-off-by: Huang Hua <huanghua78@msn.com>