Make version service return unavailable on standby masters
### What changes are proposed in this pull request?

Make version service return unavailable on standby masters

### Why are the changes needed?

In this https://github.com/Alluxio/alluxio/pull/16839 PR, we adds the
capability to run grpc services on standby masters. However, if one uses
an old alluxio client and connects to the new master with standby master
enabled, it will get some errors. This PR is used to address the issue
to make the change backward compatible.

### Does this PR introduce any user facing changes?

N/A

pr-link: Alluxio/alluxio#16854
change-id: cid-6c82bce1b1d6cb2649658ec09f0f8ed4e243917a