Add allowOnStandbyMasters option for version grpc endpoint
### What changes are proposed in this pull request?

Add a rejectOnStandbyMasters on version grpc endpoint

### Why are the changes needed?

We added a PR to make standby master return unavailable on this version
service endpoint
https://github.com/Alluxio/alluxio/pull/16854

However, in addition to the polling master inquire client, the
AbstractMasterClient also needs this endpoint.

https://github.com/Alluxio/alluxio/blob/master/core/common/src/main/java/alluxio/AbstractClient.java#L172
When a client makes a request to standby master, such check will
constantly fail and resulted in failures.

So we want the logic done in #16854 only applies to
PollingMasterInquireClient and hence we added this boolean field to
bypass the check.

### Does this PR introduce any user facing changes?

N/A

pr-link: Alluxio/alluxio#16890
change-id: cid-379f8af78250d2a230bceff9d7ea75739979e198