Use path's hash as PageId of PagingCache on worker and client
PageId of PagingCache is used as directory name on local file system.
This is expected to be a valid file name string.
Apparently a Alluxio or UFS path may contain some special charactors.
Also it may exceed file name length limit.
So, we introduce a new hash function to generate such string as dir
name. Currently SHA256 is preferred as its hash function.

Signed-off-by: Huang Hua <huanghua78@msn.com>

### What changes are proposed in this pull request?

A new AlluxioUri.hash() is introduced to generate hash for an Alluxio
path and then use it as PageId of PagingStore

### Why are the changes needed?

PageId needs a dir-name friendly string

### Does this PR introduce any user facing changes?

N/A

pr-link: Alluxio/alluxio#16863
change-id: cid-86d1be2fa86f9d73d7621573f90352c2503cff44