Fix the bug that frequent refreshing of worker info reduces throughput
### What changes are proposed in this pull request?

Fix bug that worker info is refreshed frequently.

### Why are the changes needed?

We found that worker information is refreshed frequently and
unnecessarily. This dramatically decreases the throughput since the hash
function in refresh operation takes much time. Therefore, we have to
reduce the refresh frequency and now we don't need to update the list of
worker addresses if they are not updated in a fixed interval.

### Does this PR introduce any user facing changes?

No, it doesn't.

pr-link: Alluxio/alluxio#16870
change-id: cid-951c7ee5fef5c0c75cb905ab4a9c5366248adee7