cherry-pick and resolve conflicts: Add WorkerInfoField test in CapacityCommandTest
As the title. If Options has a WorkerInfoField which POptions does not
have, or if POptions has a WorkerInfoField which Options does not have,
this test will be failed. The latest test version can not check these
two error, as shown in the PR below.

[Add missing variant in gRPC
WorkerInfoField](https://github.com/Alluxio/alluxio/pull/16457)

No.

pr-link: Alluxio/alluxio#16507
change-id: cid-23fa04dbd9b0402bff37ddb8ae42f5ec6e18f719