Fix head and tail commands read less than expected
### What changes are proposed in this pull request?

Fix a bug where `fs head` and `fs tail` output less data than it is
expected to.

### Why are the changes needed?

The code is buggy: it only calls `read` once, and does not check if the
returned number of bytes read is equal to the total number of bytes to
read as specified by the cli option.

Compare with the `cat` command:

https://github.com/Alluxio/alluxio/blob/73f3ce83c8a3ef77ac3eebb4579bb7d412784ec9/shell/src/main/java/alluxio/cli/fs/command/CatCommand.java#L57-L63

### Does this PR introduce any user facing changes?

No.

pr-link: Alluxio/alluxio#16928
change-id: cid-86b76a3444fa9efe2cd63a4b42a42e4f62b8f21b