Fix metadata sync behavior when descendant type is NONE
### What changes are proposed in this pull request?

When the metadata sync descendant type is NONE, stop loading the
children of the sync root.

If a metadata sync is trigged by a GetStatus() call on a directory
Previous behavior: The directory itself, as well as all its sub
directories in the inode store will be synced.
New behavior: ONLY the directory itself will be loaded.

### Why are the changes needed?

This PR addresses https://github.com/Alluxio/alluxio/issues/16922.
The incorrect metadata sync behavior on GetStatus for a directory loads
more children of the directory than expected and put a lot of pressure
on UFS side.

### Does this PR introduce any user facing changes?

Yes. The metadata sync behavior has been changed. See the comment above.
The previous behavior was actually wrong and we added a hidden feature
flag to allow customers to fallback.

pr-link: Alluxio/alluxio#16935
change-id: cid-2a5a2b4959422ecff74149881e400659d07c2163