Fix build error when revision is shorter than 8 chars
It will through ArrayOutOfBoundException if we build alluxio in a node
which have no git installed, and the `VERSION` would be shorter than 8,
even empty string, so substring(8) cannot work anymore.

This PR check there are more than 8 chars in the `VERSION` first,
otherwise, do not cut the `VERSION` string.

pr-link: Alluxio/alluxio#16888
change-id: cid-4e02cc9214317d86bba9d00a6121c5f013dd3255