Merge journal context in LostFileDetector
### What changes are proposed in this pull request?

`LostFileDetector` checks all files that are possibly LOST in Alluxio,
meaning the data only resides in Alluxio (not PERSISTED) and now cannot
be found(no block found).

We used to create one JournalContext when updating each file's status.
In this PR we change that to batching, to avoid too many synchronous
flushing. This is consistent with
https://github.com/Alluxio/alluxio/pull/16529

It is fine to batch because:
1. It's relatively low impact if we lose entries marking inodes as LOST.
Losing a file is HIGH impact, while not identifying a lost file is LOW
impact. That identification is asynchronous and can be turned off
anyways.
2. Even if an inode is marked LOST, we don't have any special
handling... The client will just try to read it the same way and results
in IOException. That means the LOST status is not important (unless
later we really give it some attention).

### Why are the changes needed?

Performance improvements

### Does this PR introduce any user facing changes?

Users shouldn't perceive this.

pr-link: Alluxio/alluxio#16971
change-id: cid-766a33e80bfd76df769d8261642aa1c7b18f99ee