Merge JournalContext for persistence
### What changes are proposed in this pull request?

`PersistenceChecker` and `PersistenceScheduler` will regularly check all
files that need persistence. Some fields in the inode is used to track
the persistence status. So PersistenceChecker and PersistentSchedule
will constantly update inodes, which creates journal entries.

Currently one JournalContext is created on processing each file, and
closed after this file is processed. This means the journal entries are
flushed once per file. In this change, we merge those flushes to:
1. Reduce frequent JournalContext usage, which acquires the state lock
2. Reduce the number of synchronous journal flushes

The idea is also consistent with
https://github.com/Alluxio/alluxio/pull/16529

The correctness is guaranteed by:
1. It is fine to lose those journal entries on async persist on
failover. If an inode status is not PERSISTED, the new primary will
attempt to persist it anyways.

### Why are the changes needed?

Mentioned above

### Does this PR introduce any user facing changes?

Users should not perceive this change.

pr-link: Alluxio/alluxio#16966
change-id: cid-cdb5750bb222a3a84fa219e5f9cda2045a34e137