Fix snapshot from follower thread leak
Cherry pick of https://github.com/Alluxio/alluxio/pull/15873

When using Ratis, the primary master does not take a snapshot itself,
instead it requests a snapshot from a follower. This pull request fixes
two issues with the current implementation of this.

1. Currently when a snapshot needs to be taken as given by
alluxio.master.journal.checkpoint.period.entries, Ratis will call
takeSnapshot() in the JournalStateMachine each time a journal entry is
committed until a new snapshot is installed. On the primary master
takeSnapshot() runs asynchronously by first requesting snapshot
information from each follower, then downloading a snapshot form one of
them if a valid snapshot is available. If no valid snapshot is available
(which is likely since all nodes take snapshots at the same log index
and it takes time to generate a snapshot) the request happens repeatedly
until one is available, but each request allocates a new GRPC
connection, eventually this may cause the master to crash or fail over
from allocating too many threads. This is fixed by having the follower
block until a valid snapshot is available before sending the reply (with
a configurable timeout).

2. Currently when a primary master sends a request to a follower to
start sending the snapshot it always expects the follower to start a new
RPC to do this, but if the follower does not do this (for example due to
any sort of failure or network issue) then the primary master will
always be waiting for this RPC and never install a new snapshot. This is
fixed by adding a timeout for the follower to start the RPC, and if the
timeout runs out, a new follower is tried, or the snapshot request
protocol starts again if none are available presently.

pr-link: Alluxio/alluxio#17073
change-id: cid-6f9652e57ccc9343bf0d25d21b87b5d6c17fa972