Fix metadata sync journalling when using UFS journal
This fixes an issue caused by
https://github.com/Alluxio/alluxio/pull/16529

That PR reduced the number of journal flushes during metadata sync using
embedded journal, but was not compatible with ufs journal as in ufs
journal each master has a different journal file.

This PR disables that change when using ufs journal.

pr-link: Alluxio/alluxio#17032
change-id: cid-20836bc9b5d64aff37e6f6f6e6ef3ec5f0e73617