Dump metrics + stacks on failover/crash
The master/worker processes will dump jstack/metrics when:
1. Process exits normally, from an error, or SIGTERM from
`alluxio-stop.sh`. (Note that segfault is abrupt and the process doesn't
have a chance to output these information.)
2. Primary master fails over to standby.

The major changes in the code are:
1. Dumping those information into files
2. CollectInfo and CollectLog commands will capture those output files
3. Change the jstack dumping to async
(https://github.com/Alluxio/alluxio/commit/9ae71902499b005852563055ae5135e97833e274
changed that to sync), so we don't block the process. The trade off is
the output is not totally consistent. That should be acceptable in
troubleshooting.

This change depends on
https://github.com/Alluxio/alluxio/commit/9ae71902499b005852563055ae5135e97833e274

Sample output files:

[alluxio-master-exit-stacks-20230315-063353.txt](https://github.com/Alluxio/alluxio/files/10980702/alluxio-master-exit-stacks-20230315-063353.txt)

[alluxio-master-exit-metrics-20230315-063353.txt](https://github.com/Alluxio/alluxio/files/10980711/alluxio-master-exit-metrics-20230315-063353.txt)

This helps persisting critical troubleshooting information because when
the system unexpectedly exits or failover, the admin doesn't have a
chance to manually poll these information before it's too late (process
already died or primacy is lost).

See above

pr-link: Alluxio/alluxio#17081
change-id: cid-5c9dcd5abb5c9731d3c20fe734d3f8d07f8fd167