Fix file system merge journal context
### What changes are proposed in this pull request?
1. do not use the FileSystemMergeJournalContext in list status operation
as well as other operations that calls the FileSystemMaster public
createJournalContext() method, because some of these callers does not
expect journals to be held in the context until it gets closed.
2. force flushing journals if too many journals in the
FileSystemMergeJournalContext
3. better observerability

### Why are the changes needed?

A PR that addressed inode operation consistency issue
https://github.com/Alluxio/alluxio/pull/17071/files caused issues on
master side and made fuse unable to connect to it.
https://github.com/Alluxio/alluxio/issues/17041

That PR introduced a FileSystemMergeJouralContext that keeps inode
operation journals in memory and merges these journal entries on the
fly. When the journal context closes, all of the jourals will be
appended to the journal writer. This journal context is supposed to use
in metadata write operations but it is also used unexpectedly in some
other places:
1. ListStatus might also generate joruanl entries (e.g. update access
time)
2. The journal context might also be used in other places outside the
file system master (e.g. update access time/backup/checkpoint etc.)
These places might generate excessive # of journal entries and all kept
in memory and might cause OOM issue.
So in the fix, we will just return a normal non-merging journal context
instead.

### Does this PR introduce any user facing changes?

N/A

pr-link: Alluxio/alluxio#17071
change-id: cid-02ed825792b57fe6fb1b1b7fcb4566ff5caabb7d