Fix dependencies conflicts thrown by trino
**PR Background - Bug description:**
[PrestoDB](https://prestodb.io/) and [TrinoDB](https://trino.io/) use
our shaded JAR in their `pom.xml`. When upgrading the version of
`alluxio-shaded-client` from `2.8.1` to `2.9.1`, we fail to compile the
codes of presto/trino. This is because there is dependency conflict.

**Solution:**
Relocate the dependencies in `shaded/client/pom.xml` to avoid deplicate
classes and files.

**TIPS:** You can execute `cd ${ALLUXIO_HOME}/shaded && mvn clean
install -Prelease -DskipTests -T 4C` to package the shaded JAR.

pr-link: Alluxio/alluxio#17129
change-id: cid-1c9e013c98369ebb206b5c5d52966780ee34944c