cherry-pick and resolve conflicts: Add worker decommission step 1
This pr is related to #13758

This change achieves below functionalities:
1. Provide a client shell command to kick out workers directly.
2. Decommissioned worker nodes are not automatically shutdown and are
not chosen for new IO requests
3. Decommissioned workers are not accepted on heartbeat or register

Note there's a delay in when clients know the workers have been
decommissioned. In the window, target worker will still receive
requests. There's also no way for the worker to come back to the
cluster. This command should only be used when this worker is going to
leave the cluster (and never come back).

Offline workers more flexible, no need to wait master detection lost
worker interval.
For instance,
We may encounter situations like we want to remove worker nodes, or we
want to offline some worker nodes to save costs, this command is useful.
Yes

pr-link: Alluxio/alluxio#14000
change-id: cid-822a1cad1b23cd381b907705c1b92bd6ac3f16d0