Fix slow S3 read
### What changes are proposed in this pull request?

Fix slow S3 read.

### Why are the changes needed?

The UFS reader uses a input stream-to-channel adapter that uses an internal buffer to transfer data from UFS input stream to output buffer. The internal buffer is small (8 KiB). With S3, the S3 stream will have to send a GetObject request every time a read call is made, and small buffers causes a lot of small reads. This is inefficient use of the S3 stream.

### Does this PR introduce any user facing changes?

No.
			pr-link: Alluxio/alluxio#17245
			change-id: cid-d2f52425a4fbb7041b40fd82ccfa767bf631c72c