Fix FileSystemMergeJournalContext related
1. Only force flushing journals when a locked inode path is closed for FileSystemMergeJournalContext
2. Fix the journal merger broken logic where updating the fingerprint for a directry will be ignored.

We used to fix the merge journal context in https://github.com/Alluxio/alluxio/pull/17071, where a regular non-merging journal context is used when listStatus() is called.

However, if a listStatus triggers a metadata sync and MASTER_FILE_SYSTEM_MERGE_INODE_JOURNALS is set to true, journals will be flushed on every lockedInodePath close during the metadata sync. This behavior leads to journals being flushed too many times and impairs the metadata sync performance.

Also we found a minor issue that when inode directory journals are merged, the fingerprint will be ignored. This is becuase inode directory journal does not have a fingerprint field.

N/A
			pr-link: Alluxio/alluxio#17251
			change-id: cid-23ea3229e6781483ab582c3f4d4c9c4f61cdf634