Add virtual blocks by splitting a file to improve Presto's parallelism
Although Dora doesn't has the Block concept any more, Presto Coordinator still depends on the block information to construct split when doing a TableScan Operator. It is not so good to regard a file as a big block because if a large file takes too much time for a Presto Task.

Therefore, this PR aims at constructing multiple virtual blocks for Presto Coordinator by splitting a file according to a fixed size.
			pr-link: Alluxio/alluxio#17297
			change-id: cid-54c36ee027ac40aaa46969120978ff5cc8e2ad02