Set Metadata cache TTL to 48 hours
This will result in longer cache life for metadata. It increases metadata operation performance. This is suitable for read-only use case.

### What changes are proposed in this pull request?

Increase TTL values for metadata cache. Increase to 48 hours. If the metadata is not used in 48 hours, they can be considered not needed in near future.

### Why are the changes needed?

For better performance.

### Does this PR introduce any user facing changes?

N/A

			pr-link: Alluxio/alluxio#17304
			change-id: cid-60041ad378b5225c077b94a49fbce526b5ddef94