Use dynamic buffering for sequential read in PositionReadFileInStream
### What changes are proposed in this pull request?

Use dynamic buffering for sequential read.

### Why are the changes needed?

It's hard to assume the consumer of a seekable, positioned readable input stream will be doing sequential reads. If the consumer randomly reads from different positions, then it would be pointless to do aggressive buffering since it would only cause read amplification and wastes time and bandwith.

Otherwise, if the read sequence is consecutive, then it can benefit from prefetching and buffering.

The idea of this PR is to first observe the read sequence, calculate the length of consecutive reads, and based on that, tries to prefetch data from the underlying reader. It keeps a sliding window of recent read calls, and use the sum of the consecutive read lengths in the window as the size of the prefetch.

With dynamic buffering, the prefetch size will gradually grow bigger if the sequence stays consecutive. It will be reset to 0 if any non-sequential read occurs.

### Does this PR introduce any user facing changes?

No.

			pr-link: Alluxio/alluxio#17306
			change-id: cid-3f0835aa1b720f94cad59d663ad684b589069744