[DOCFIX] Add docs for rolling upgrade and decommission worker
This is the doc update from a few recent commits:
1. https://github.com/Alluxio/alluxio/commit/aee3c5cb960b2e472f4fcdb86e731ca3b91d2f8c
2. https://github.com/Alluxio/alluxio/commit/9a4e154e7293667572fec25c63136462fb58a345
3. https://github.com/Alluxio/alluxio/commit/dbac084c1e1d89c1e9253da036746e2df9e1ef14
4. https://github.com/Alluxio/alluxio/commit/8da59539201cff376c33256846bd565d976a148e
5. https://github.com/Alluxio/alluxio/commit/141ee0e567a4c6c60907f85ea66b828704bd761d
6. https://github.com/Alluxio/alluxio/commit/26257e6f35c57d937f6f5bdee831facc2fcf7e1a

In short, a few improvements have been added that the admin can now see the version/revision of the running Alluxio components in the cluster. A few commands were also added to help with gracefully decommissioning/restarting workers in the cluster.

This doc change helps users understand how to utilize those features.

			pr-link: Alluxio/alluxio#17381
			change-id: cid-1d45c1750f84ac94d3fa79b9bbe0f0f328f7cb45