Fix startup master health check
A workaround to fix the master health check issue, related to #17382 which is caused by a recent change that starts grpc servers on standby masters.
Follower masters have a gRPC server always active. When the `MasterHealthCheckClient` connects to the gRPC server, it may find that some services aren't registered (NotFound) rather than unavailable. 
This manifests in a bug where `bin/alluxio-start.sh master` performs a health check after starting a master and incorrectly reports the master as unhealthy. 
			pr-link: Alluxio/alluxio#17383
			change-id: cid-38f5babf01a3597099a821ec3be61f325630c098