cherry-pick and resolve conflicts: Make the client poll the master concurrently
Make the client poll the master concurrently.

Currently, we find the active master node by polling each node in series. If a node did not respond due to network or other issues, this may take a lot of time and affect service availability. Make the client poll the master concurrently can alleviate this problem.

			pr-link: Alluxio/alluxio#17186
			change-id: cid-d1ef20b596deda02375324978adb23e6f8ce8dce