Make the client poll the master concurrently
### What changes are proposed in this pull request?

Make the client poll the master concurrently.

### Why are the changes needed?
Currently, we find the active master node by polling each node in series. If a node did not respond due to network or other issues, this may take a lot of time and affect service availability. Make the client poll the master concurrently can alleviate this problem.

### Does this PR introduce any user facing changes?

			pr-link: Alluxio/alluxio#17186
			change-id: cid-d1ef20b596deda02375324978adb23e6f8ce8dce