Remove hardcoded VERSION in libexec/alluxio-config.sh
previously the VERSION env variable was defined in libexec/alluxio-config.sh.
the variable exists to properly locate various files bundled as part of the compiled alluxio package, such as when bundling as a tarball, because the pom.xml file is included.
however this lead to having two sources of defining the version string, one in pom.xml and the other in alluxio-config.sh.
they happened to be usually in sync because update-versions.sh will replace the version string in both of them simultaneously.

after this change, the version string is printed out when compiling the repo, written to a generated version.sh file
the file is sourced by alluxio-config.sh to maintain backcompat and is included in the tarball packaging
this results in the single source of truth being the pom.xml in the repo.
libexec/version.sh only serves as the source of truth when pom.xml is not present, within a compiled alluxio package