Use local conf to create filesystem for s3proxy
### What changes are proposed in this pull request?

When the filesystem establishes a connection with the master, the configuration of the filesystem will be overridden by the master's configuration, resulting in changes. Therefore, we should not use the configuration of the filesystem anymore, but instead use the configuration of the local client.

### Why are the changes needed?

It is very common for the master configuration and client configuration to be different. We should avoid the newly created filesystem from being influenced by the server.

### Does this PR introduce any user facing changes?

Please list the user-facing changes introduced by your change, including
  1. change in user-facing APIs
  2. addition or removal of property keys
  3. webui

			pr-link: Alluxio/alluxio#17439
			change-id: cid-1af4fd37ee8c52cd32552818989dce557256b228