Remove hardcoded VERSION in libexec/alluxio-config.sh
previously the VERSION env variable was defined in libexec/alluxio-config.sh. the variable exists to properly locate various files bundled as part of the compiled alluxio package, such as when bundling as a tarball, because the pom.xml file is not included. however this led to having two sources of defining the version string, one in pom.xml and the other in alluxio-config.sh. they happened to be usually in sync because update-versions.sh will replace the version string in both of them simultaneously.

after this change, the version string is printed out when compiling the repo, written to a generated version.sh file. this file is sourced by alluxio-config.sh to maintain backcompat and is included in the tarball packaging. this results in the single source of truth:
- within the repo, the pom.xml defines the version and subsequently generates libexec/version.sh
- outside of the repo when pom.xml is not present, such as in the packaged tarball, libexec/version.sh defines the version

			pr-link: Alluxio/alluxio#17431
			change-id: cid-610e839c29646c4f524ab8b7aa30fe0ceaa04a10