Resolve the UFS leak in copy job
### What changes are proposed in this pull request?

This PR resolves the UFS leak problem in copy job. By making the method public in UfsManager.
### Why are the changes needed?

Since we don't have mountID in Dora, so we need a new way to identify which UFS to use. Currently, I'm following the original design which uses Scheme, Authority & Properties as the key. Later on, we should also use the token given by the client to identify the UFS.

### Does this PR introduce any user facing changes?

na

			pr-link: Alluxio/alluxio#17360
			change-id: cid-6482b34834d07496a61b29bcef7e63a6de8a9d5a