Remove mount check when starting alluxio if users do not intend to mount
### What changes are proposed in this pull request?

Remove mount check when starting alluxio if users do not intend to mount

### Why are the changes needed?

When the user does not want to mount and starts alluxio directly, the user can enter `./bin/alluxio-start.sh local` or `./bin/alluxio-start.sh all`. The old code will not be able to start alluxio successfully, it will refuse to start or the AlluxioWorker will fail to start. This PR fixes the issue.

![截屏2023-06-08 17 25 44](https://github.com/Alluxio/alluxio/assets/132879614/53cd6cb6-19e0-424b-8ff2-8ba49c8f0cb6)

The reason for deleting the mount check part of the code is:
We found that the mount check is only started when the user enters `./bin/alluxio-start.sh worker` or `./bin/alluxio-start.sh local`. However, if the user intends to use alluxio after mounting, the user will use `alluxio-mount.sh` for the mount operation. If the mount fails, an error will be reported, and `alluxio-start.sh` will not be executed. If alluxio is started directly without mounting, there is no need to perform mount check. Therefore, under the logic of our current `main` branch, this mount check is redundant and will cause alluxio to fail to start. This problem can be corrected after deletion.

As for `alluxio-start.sh local SudoMount` involves mount check, execute `alluxio-start.sh local SudoMount` to check whether it is mounted before. If it has been mounted before, the mount command will not be executed. If it has not been mounted before, mount it load disk. But the check mount here is redundant. This function is to force mounting before using alluxio, which does not meet the purpose of "use alluxio directly without mounting". After removing the check mount function, it does not affect the use of this command.

![截屏2023-06-14 11 57 35](https://github.com/Alluxio/alluxio/assets/132879614/4cc56dfa-8674-499f-a263-c0d672de5b21)

### Does this PR introduce any user facing changes?

No

			pr-link: Alluxio/alluxio#17603
			change-id: cid-e67fe1697e4c75fbf0b565a00675f57f6136cc31