Delete unrecginized page file and handle empty result from getPageInfo()
### What changes are proposed in this pull request?
- delete recognized page file, including TEMP pages
- handle empty result from getPageInfo(), which returns Optional.empty() if page is not recognized

### Why are the changes needed?

When Worker (re-)starts, worker scans the StoreDir to establish the metadata for PagingStore.
If some page files are not recognized, the original solution is to delete all metadata and all existing page files.
One special case is page files in TEMP dir. These TEMP page files are not recognized as normal page files.
And also, we don't know if these TEMP page files are valid or complete. So, we have to delete those unrecognized page files (including TEMP page files).

When this happens, we also need to handle the Optional.empty result from getPageInfo().

### Does this PR introduce any user facing changes?
N/A

			pr-link: Alluxio/alluxio#17593
			change-id: cid-8889577f492e46470411eedeed80dfaa613a28aa