Simplify `elect` command
After upgrading Apache Ratis to version 2.5.1 (#17571), it is possible to transfer leadership between two masters of equal priority. This PR removes the need to update priorities twice when using the `elect` command.
In addition, this PR reworks the inner working of the command to make them thread safe: the new logic ensures that the transfer id is selected and reserved atomically. Moreover, it ensures that the transfer id is returned to the client before initiating the transfer of leadership. This is important because the leader receiving this RPC will be demoted and close its RPC server, potentially propagating an error to the client erroneously. 
			pr-link: Alluxio/alluxio#17598
			change-id: cid-1de7738fcf7f96b1e1b1ca509c04f914f835c447